// src/app/auth/types/authState.interface.ts
import { CurrentUserInterface } from 'src/app/shared/types/currentUser.interface';
import { BackendErrorsInerface } from 'src/app/shared/types/backendErrors.interface';

export interface AuthStateInterface {
  isSubmitting: boolean;
  currentUser: CurrentUserInterface | null;
  isLoggedIn: boolean | null;
  validationBackendErrors: BackendErrorsInerface | null;
  isLoading: boolean;
}